voice_text/requirements.txt
openai-whisper==20231117
pygame==2.5.2
gtts==2.5.1
pyaudio==0.2.14
python-dotenv==1.0.1
scipy==1.13.1
ffmpeg-python==0.2.0
#psutil==5.9.5
langchain == 0.1.8
asyncio == 3.4.3
sse-starlette == 1.6.1
pydantic == 1.10.9
uvicorn == 0.23.2
langchain == 0.1.8
tiktoken 
pypdf == 3.16.2
#ctransformers == 0.2.27
langchain_community == 0.0.21
PyPDF2 == 3.0.1
python-multipart == 0.0.9
docx2txt == 0.8
accelerate == 0.29.0
pdf2image == 1.17.0
unstructured 
pdfminer.six
opencv-python
huggingface_hub == 0.19.4
openai == 1.30.0
python-dotenv
rank_bm25
pandas == 2.2.0
streamlit == 1.35.0
langchain-openai
fastapi == 0.99.1